server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain-openai==0.1.0
langchain-community==0.0.20
langgraph==0.0.40
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
